fix for drawBiggestSqure -> changed floats to double

+++ randomWalk/CPU/rectHost.h
@@ -3,7 +3,7 @@
 
 #include <iostream>
 
-typedef float floatingPoint;
+typedef double floatingPoint;
 
 struct point 
 {
